Resolve the issue that the LINQ query can't be translated and will be evaluated locally.

+++ src/Kendo.DynamicLinqCore.Tests/CustomJsonSerializerOptions.cs
@@ -40,7 +40,7 @@ namespace Kendo.DynamicLinqCore.Tests
                     return l;
                 }
 
-                return reader.GetDecimal();
+                return reader.GetDouble();
             }
 
             if (reader.TokenType == JsonTokenType.String)
